Tests/LibWeb/Text/expected/css/at-property-initial-value-background-color.txt

Computed background-color: rgb(255, 0, 0)
